+++ src/index.tsx
@@ -1,15 +1,4 @@
-import React, { type PropsWithChildren } from 'react';
-import { Text as RNText } from 'react-native';
+import { Circle } from './components/Circle';
+import { Typography } from './components/Typography';
 
-interface Props {
-  color?: string;
-  fontSize: number;
-}
-
-export const Typography: React.FC<PropsWithChildren<Props>> = ({
-  color = 'black',
-  fontSize,
-  children,
-}) => {
-  return <RNText style={{ color, fontSize }}>{children}</RNText>;
-};
+export { Circle, Typography };
